Ignore KuduClient, lint DeploymentSwap and ErrorData (#269)
* Ignore KuduClient, lint DeploymentSwap and ErrorData
